requirements.txt
amqp==5.0.9
asgiref==3.5.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.3
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.0
djangorestframework==3.13.1
kombu==5.2.3
prompt-toolkit==3.0.24
psycopg2-binary==2.9.3
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
vine==5.0.0
wcwidth==0.2.5
